requirements.txt
# Core Backend & API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
redis==5.0.1

# HTTP & Web Crawling
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.40.0
aiohttp==3.9.1

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
joblib==1.3.2

# Environment & Config
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Web requests & parsing
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3

# For future async operations (we'll use later)
aiofiles==23.2.1